requirements.txt
anyio==3.6.2
async-timeout==4.0.2
autopep8==2.0.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
dnspython==2.2.1
fastapi==0.87.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.3
PyMySQL==1.0.2
python-dotenv==0.21.0
redis==4.5.4
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
Werkzeug==2.2.2
